Fix for Driver error, when more than 2 populate are done on a model
A async.each loop iterates over the 2nd populate onwards in a find. The first populate is always a native join. Since async.each goes in parallel, the same connection is used to execute the additional populates which result in 'Requests can only be made in the LoggedIn state, not the SentClientRequest state' error.

Updated the code to use async.eachSeries so that only one query is executed at a time.